runScripts/requirements.txt
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
lazy_loader==0.4
llvmlite==0.44.0
MarkupSafe==3.0.2
narwhals==1.43.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
opencv-python-headless==4.11.0.86
packaging==24.2
pandas==2.3.0
pillow==11.2.1
platformdirs==4.3.8
pooch==1.8.2
protobuf==6.31.1
pyarrow==20.0.0
pydeck==0.9.1
PyMatting==1.1.14
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
rembg==2.0.66
requests==2.32.4
rpds-py==0.25.1
scikit-image==0.25.2
scipy==1.15.3
six==1.17.0
smmap==5.0.2
streamlit==1.45.1
tenacity==9.1.2
tifffile==2025.6.11
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
